requirements.txt
jira==3.8.0
python-dotenv==1.0.1